Updated init.sh to configure auth.json automatically
When the global auth.json file is present and webroot/auth.json does not exist, the init.sh script will now detect this and copy credentials for repo.magento.com into the webroot/auth.json file to meet the required file check. In an ideal world, this file will be created by the developer with marketplace credentials specific to the project as will be needed to install project specific extensions, but this will streamline intial setup for those with global auth credentials configured.

A notice will be output to stderr when the file is autoconfigured:

> NOTICE: Configuring ./webroot/auth.json with global credentials for repo.magento.com